+++ impeller/golden_tests/golden_playground_test_mac.cc
@@ -69,22 +69,6 @@ static const std::vector<std::string> kSkipTests = {
     "impeller_Play_AiksTest_CanRenderClippedRuntimeEffects_Vulkan",
 };
 
-// clang-format off
-static const std::vector<std::string> kVulkanDenyValidationTests = {
-    "impeller_Play_GaussianBlurFilterContentsTest_RenderCoverageMatchesGetCoverageRotated_Vulkan",
-    "impeller_Play_SceneTest_FlutterLogo_Vulkan",
-    "impeller_Play_SceneTest_CuboidUnlit_Vulkan",
-    "impeller_Play_RuntimeStageTest_CanCreatePipelineFromRuntimeStage_Vulkan",
-    "impeller_Play_RuntimeEffectSetsRightSizeWhenUniformIsStruct_Vulkan",
-    "impeller_Play_RuntimeEffectCanSuccessfullyRender_Vulkan",
-    "impeller_Play_RuntimeEffect_Vulkan",
-    "impeller_Play_EntityTest_RuntimeStageTest_CanCreatePipelineFromRuntimeStage_Vulkan",
-    "impeller_Play_EntityTest_RuntimeEffectSetsRightSizeWhenUniformIsStruct_Vulkan",
-    "impeller_Play_EntityTest_RuntimeEffectCanSuccessfullyRender_Vulkan",
-   "impeller_Play_EntityTest_RuntimeEffect_Vulkan",
-};
-// clang-format on
-
 namespace {
 std::string GetTestName() {
   std::string suite_name =
