+++ impeller/playground/playground.h
@@ -32,6 +32,22 @@ enum class PlaygroundBackend {
   kVulkan,
 };
 
+// TODO(https://github.com/flutter/flutter/issues/145039)
+// clang-format off
+static const std::vector<std::string> kVulkanDenyValidationTests = {
+  "impeller_Play_SceneTest_FlutterLogo_Vulkan",
+  "impeller_Play_SceneTest_CuboidUnlit_Vulkan",
+  "impeller_Play_RuntimeStageTest_CanCreatePipelineFromRuntimeStage_Vulkan",
+  "impeller_Play_RuntimeEffectSetsRightSizeWhenUniformIsStruct_Vulkan",
+  "impeller_Play_RuntimeEffectCanSuccessfullyRender_Vulkan",
+  "impeller_Play_RuntimeEffect_Vulkan",
+  "impeller_Play_EntityTest_RuntimeStageTest_CanCreatePipelineFromRuntimeStage_Vulkan",
+  "impeller_Play_EntityTest_RuntimeEffectSetsRightSizeWhenUniformIsStruct_Vulkan",
+  "impeller_Play_EntityTest_RuntimeEffectCanSuccessfullyRender_Vulkan",
+  "impeller_Play_EntityTest_RuntimeEffect_Vulkan",
+};
+// clang-format on
+
 constexpr inline RuntimeStageBackend PlaygroundBackendToRuntimeStageBackend(
     PlaygroundBackend backend) {
   switch (backend) {
