+++ tools/gn
@@ -1256,18 +1256,18 @@ def parse_args(args):
   )
 
   parser.add_argument(
+      '--enable-impeller-opengles',
+      default=True,
+      action='store_true',
+      help='Enable the Impeller OpenGL ES backend.'
+  )
+
+  parser.add_argument(
       '--prebuilt-impellerc',
       default=None,
       type=str,
       help='Absolute path to a prebuilt impellerc. ' +
       'Do not use this outside of CI or with impellerc from a different engine version.'
-  )
-
-  parser.add_argument(
-      '--enable-impeller-opengles',
-      default=True,
-      action='store_true',
-      help='Enable the Impeller OpenGL ES backend.'
   )
 
   parser.add_argument(
